Add object match endpoint
match will take each parameter in document contents and use operator to match the records.
it will also take return type to use case define what needs to be returned.
token: generate a token if records matched : this can be mainly used for application want to have a token to control the login

matchObject is not used to get document, and it will take the return type as long as there is one match under the condition.